In this massive commit, I've redesigned from scratch the helper architecture to support some new features: - Flexible data communication from the helper - An API to execute multiple actions with a single request - An elegant way to receive signals from the helper, without   making Action a QObject.
I've added an ActionWatcher class. The only way to obtain an
instance of it is with the member watcher() of an Action object.
It'll return an object that you can use to connect to the signals
from the helper.

To support this, I had to include a way to know which action
a signal was about, but this has not much sense if the helper 
knows only about one action at the time, so I've added the
support for multiple actions. With one call to 
Action::executeActions(), one can execute a list of actions
in a single shot. Currently, the helper executes them sequentially,
but now, every signal that it sends has the knowledge of which
action it's about. This is how the HelperProxy decide which
ActionWatcher should emit the signal to the user.

Single asynchronous calls are possibile with the convenience
method executeAsync(). It has two parameters to simply connect
the actionPerformed() slot without having to use the watcher
if you don't need it. If any of the two parameters are NULL,
the action is performed but no signal is delivered (if you
don't use a watcher yourself, of course)

Single synchronous calls are still possible with execute(), which
behaves exactly as before.


svn path=/trunk/playground/base/libkauth/example/; revision=991906