New KWin example script: demands attention on current desktop
The rational behind this script is to have only windows on current
desktop as demanding attention. Restricting the demands attention
state to the current desktop means for example that the tasks applet
won't show windows started on different desktops.

The script keeps track of the windows where it removed demands
attention and restores the state when the window's desktop becomes
active.

CCBUG: 231898